+++ R/assess_SCA_RWM.R
@@ -5,7 +5,8 @@
 #' Set the variance (\code{start$tau_M}) to a small value (0.001) in order to fix M for all years, which is functionally equivalent to \link{SCA}.
 #' 
 #' @inheritParams SCA 
-#' @param refyear An expression for the year for which M is used to report MSY reference points. By default, terminal year.
+#' @param refyear An expression for the year for which M is used to report MSY reference points. By default, terminal year. IF multiple
+#' years are provided, then the mean M over the specified time period is used.
 #' @param M_bounds A numeric vector of length 2 to indicate the minimum and maximum M in the random walk as a proportion of the starting M
 #' (\code{start$M}). The default min and max are 75\% and 125\%, respectively.
 #' @details