+++ NEWS.md
@@ -2,6 +2,7 @@ The latest release of the SAMtool package is available on [CRAN](https://CRAN.R-
 
 ## SAMtool 1.2.4
 - Update `Shortcut` indexing to align year of assessment with projection. An MP using the `Perfect` assessment and `HCR_MSY` annually will produce F = FMSY in the OM.
+- Stabilize `VPA` when the catch-at-age in the plusgroup and plusgroup-1 is very small.
 
 ## SAMtool 1.2.3
 - `RCM` will check age and length comp data for NA's and replaces with zero